cloud_function/requirements.txt
requests>=2.25.1
google-cloud-firestore>=2.0.0
